Move useMCMCMonitor
