+++ src/components/AutocorrelationPlot.tsx
@@ -1,5 +1,5 @@
 import { FunctionComponent, useMemo } from "react";
-import { useMCMCMonitor } from "../util/useMCMCMonitor";
+import { useMCMCMonitor } from "../MCMCMonitorDataManager/useMCMCMonitor";
 import AutocorrelationPlotWidget from "./AutocorrelationPlotWidget";
 
 type Props = {
