+++ src/components/RunInfoTab.tsx
@@ -1,6 +1,6 @@
 import { Grid } from "@mui/material";
 import { FunctionComponent, useMemo } from "react";
-import { useMCMCMonitor } from "../util/useMCMCMonitor";
+import { useMCMCMonitor } from "../MCMCMonitorDataManager/useMCMCMonitor";
 
 type Props = {
 	runId: string
