+++ src/components/SequenceHistogram.tsx
@@ -1,5 +1,5 @@
 import { FunctionComponent, useMemo } from "react";
-import { useMCMCMonitor } from "../util/useMCMCMonitor";
+import { useMCMCMonitor } from "../MCMCMonitorDataManager/useMCMCMonitor";
 import SequenceHistogramWidget from "./SequenceHistogramWidget";
 
 type Props = {
