+++ src/components/SequenceScatterplot.tsx
@@ -1,5 +1,5 @@
 import { FunctionComponent, useMemo } from "react";
-import { useMCMCMonitor } from "../util/useMCMCMonitor";
+import { useMCMCMonitor } from "../MCMCMonitorDataManager/useMCMCMonitor";
 import { applyDrawRange } from "./SequenceHistogram";
 import SequenceScatterplotWidget, { ScatterplotSequence } from "./SequenceScatterplotWidget";
