+++ src/components/TablesTab/MainTable.tsx
@@ -1,6 +1,6 @@
 import { Table, TableBody, TableCell, TableHead, TableRow } from "@mui/material";
 import { FunctionComponent, useMemo } from "react";
-import { useMCMCMonitor } from "../../util/useMCMCMonitor";
+import { useMCMCMonitor } from "../../MCMCMonitorDataManager/useMCMCMonitor";
 
 type Props = any
